Refactored Assxml exporter and dump xml writer.
- Moved AssxmlExporter serialization code in a new file AssxmlFileWriter.cpp/h
- Added new files to CMakeLists.txt
- Replaced string format specifiers %i by %u to match argument type.
- Made a few changes in WriteDumb.cpp to call the new DumpSceneToAssxml function.
